Use getParserName
