+++ src/language-markdown/embed.js
@@ -1,7 +1,6 @@
 "use strict";
 
 const util = require("../common/util");
-const support = require("../main/support");
 const {
   builders: { hardline, literalline, concat, markAsRoot },
   utils: { mapDoc },
@@ -15,7 +14,7 @@ function embed(path, print, textToDoc, options) {
     // only look for the first string so as to support [markdown-preview-enhanced](https://shd101wyy.github.io/markdown-preview-enhanced/#/code-chunk)
     const langMatch = node.lang.match(/^[\w-]+/);
     const lang = langMatch ? langMatch[0] : "";
-    const parser = getParserName(lang);
+    const parser = util.getParserName(lang, options);
     if (parser) {
       const styleUnit = options.__inJsTemplate ? "~" : "`";
       const style = styleUnit.repeat(
@@ -64,22 +63,6 @@ function embed(path, print, textToDoc, options) {
   }
 
   return null;
-
-  function getParserName(lang) {
-    const supportInfo = support.getSupportInfo({ plugins: options.plugins });
-    const language = supportInfo.languages.find(
-      (language) =>
-        language.name.toLowerCase() === lang ||
-        (language.aliases && language.aliases.includes(lang)) ||
-        (language.extensions &&
-          language.extensions.find((ext) => ext === `.${lang}`))
-    );
-    if (language) {
-      return language.parsers[0];
-    }
-
-    return null;
-  }
 
   function replaceNewlinesWithLiterallines(doc) {
     return mapDoc(doc, (currentDoc) =>
